feat(types): Add helper type to easy add HTML props
Also ... smal rearrange stuff and delete stuff we don't use anyway.
