--- packages/types/src/array.ts
@@ -1,45 +1,0 @@
-import { Mutable } from './object';
-
-/**
- * Get values of an `array` as literals.
- * *Note:* To make this work you should use `as const`.
- */
-export type ValueOf<T extends any[] | Readonly<any[]>> = Mutable<T>[number];
-
-/**
- * Flatten an `array`. This will land in TS soon.
- * See: https://github.com/microsoft/TypeScript/pull/32131
- */
-// TODO: This can not be parsed by latest eslint .... (as of 04-02-2021)
-// export type Flat<A, D extends number> = {
-//   '1': A;
-//   '0': A extends readonly [infer T]
-//     ? Flat<
-//         T,
-//         [
-//           -1,
-//           0,
-//           1,
-//           2,
-//           3,
-//           4,
-//           5,
-//           6,
-//           7,
-//           8,
-//           9,
-//           10,
-//           11,
-//           12,
-//           13,
-//           14,
-//           15,
-//           16,
-//           17,
-//           18,
-//           19,
-//           20
-//         ][D]
-//       >
-//     : A;
-// }[D extends -1 ? '1' : '0'];
