--- packages/types/src/object.ts
@@ -1,4 +1,0 @@
-/**
- * Make an object or array mutable.
- */
-export type Mutable<T> = { -readonly [P in keyof T]: T[P] };
